+++ apps/wizard/app_pages/dashboard/app.py
@@ -3,17 +3,19 @@
 import subprocess
 from datetime import datetime
 
-import pandas as pd
 import streamlit as st
-from st_aggrid import AgGrid, GridUpdateMode, JsCode
-from st_aggrid.grid_options_builder import GridOptionsBuilder
 from structlog import get_logger
 
-from apps.step_update.cli import StepUpdater, UpdateState
-from apps.wizard.app_pages.dashboard.utils import NON_UPDATEABLE_IDENTIFIERS, _create_html_button
+from apps.wizard.app_pages.dashboard.agrid import make_agrid
+from apps.wizard.app_pages.dashboard.utils import (
+    NON_UPDATEABLE_IDENTIFIERS,
+    _create_html_button,
+    check_db,
+    load_steps_df,
+    load_steps_df_to_display,
+)
 from apps.wizard.utils.components import st_horizontal
 from etl.config import OWID_ENV
-from etl.db import can_connect
 
 st.set_page_config(
     page_title="Wizard: ETL Dashboard",
@@ -78,119 +80,15 @@ with st_horizontal():
 ########################################
 st.markdown("### Steps table")
 
-if not can_connect():
-    st.error("Unable to connect to grapher DB.")
-
-
-@st.cache_data
-def load_steps_df(reload_key: int) -> pd.DataFrame:
-    """Generate and load the steps dataframe.
-
-    This is just done once, at the beginning.
-    """
-    # Ensure that the function is re-run when the reload_key changes.
-    _ = reload_key
-
-    # Load steps dataframe.
-    steps_df = StepUpdater().steps_df
-
-    # Fix some columns.
-    steps_df["full_path_to_script"] = steps_df["full_path_to_script"].fillna("").astype(str)
-    steps_df["dag_file_path"] = steps_df["dag_file_path"].fillna("").astype(str)
-
-    # For convenience, convert days to an arbitrarily big number.
-    # Otherwise when sorting, nans are placed before negative numbers, and hence it's not easy to first see steps that
-    # need to be updated more urgently.
-    steps_df["days_to_update"] = steps_df["days_to_update"].fillna("9999")
-
-    # For convenience, combine dataset name and url in a single column.
-    # This will be useful when creating cells with the name of the dataset as a clickable link.
-    # In principle, one can access different columns of the dataframe with UrlCellRenderer
-    # (and then hide db_dataset_id column), however, then using "group by" fails.
-    # So this is a workaround to allows to have both clickable cells with names, and "group by".
-    steps_df["db_dataset_name_and_url"] = [
-        f"[{row['db_dataset_name']}]({OWID_ENV.dataset_admin_site(int(row['db_dataset_id']))})"
-        if row["db_dataset_name"]
-        else None
-        for row in steps_df.to_dict(orient="records")
-    ]
-
-    steps_df = steps_df.drop(columns=["db_dataset_name", "db_dataset_id"], errors="raise")
-
-    return steps_df
-
-
-@st.cache_data
-def load_steps_df_to_display(show_all_channels: bool, reload_key: int) -> pd.DataFrame:
-    """Load the steps dataframe, and filter it according to the user's choice."""
-    # Load all data
-    df = load_steps_df(reload_key=reload_key)
-
-    # If toggle is not shown, pre-filter the DataFrame to show only rows where "channel" equals "grapher"
-    if not show_all_channels:
-        df = df[df["channel"].isin(["grapher", "explorers"])]
-
-    # Sort displayed data conveniently.
-    df = df.sort_values(
-        by=["days_to_update", "n_chart_views_365d", "n_charts", "kind", "version"],
-        na_position="last",
-        ascending=[True, False, False, False, True],
-    )
-
-    # Prepare dataframe to be displayed in the dashboard.
-    df = df[
-        [
-            "step",
-            "db_dataset_name_and_url",
-            "days_to_update",
-            "update_state",
-            "n_charts",
-            # "n_chart_views_7d",
-            "n_chart_views_365d",
-            "update_period_days",
-            "date_of_next_update",
-            "namespace",
-            "version",
-            "channel",
-            "name",
-            "kind",
-            "dag_file_name",
-            "n_versions",
-            # "state",
-            "full_path_to_script",
-            "dag_file_path",
-            # "versions",
-            # "role",
-            # "all_usages",
-            # "direct_usages",
-            # "all_chart_ids",
-            # "all_chart_slugs",
-            # "all_chart_views_7d",
-            # "all_chart_views_365d",
-            # "all_active_dependencies",
-            # "all_active_usages",
-            # "direct_dependencies",
-            # "chart_ids",
-            # "same_steps_forward",
-            # "all_dependencies",
-            # "same_steps_all",
-            # "same_steps_latest",
-            # "latest_version",
-            # "identifier",
-            # "same_steps_backward",
-            # "n_newer_versions",
-            # "db_archived",
-            # "db_private",
-        ]
-    ]
-    return df
-
+# Check if the database is accessible.
+check_db()
 
 # Streamlit UI to let users toggle the filter
 show_all_channels = not st.toggle("Select only grapher and explorer steps", True)
 
 # Load the steps dataframe.
-steps_df = load_steps_df(reload_key=st.session_state["reload_key"])
+with st.spinner("Loading steps details from ETL and DB..."):
+    steps_df = load_steps_df(reload_key=st.session_state["reload_key"])
 
 
 ########################################
@@ -199,195 +97,8 @@ steps_df = load_steps_df(reload_key=st.session_state["reload_key"])
 # Get only columns to be shown
 steps_df_display = load_steps_df_to_display(show_all_channels, reload_key=st.session_state["reload_key"])
 
-# Define the options of the main grid table with pagination.
-gb = GridOptionsBuilder.from_dataframe(steps_df_display)
-gb.configure_grid_options(domLayout="autoHeight", enableCellTextSelection=True)
-gb.configure_selection(
-    selection_mode="multiple",
-    use_checkbox=True,
-    rowMultiSelectWithClick=True,
-    suppressRowDeselection=False,
-    groupSelectsChildren=True,
-    groupSelectsFiltered=True,
-)
-gb.configure_default_column(editable=False, groupable=True, sortable=True, filterable=True, resizable=True)
-gb.configure_column("step", headerName="Step", width=500, headerTooltip="Step URI, as it appears in the ETL DAG.")
-gb.configure_column(
-    "channel", headerName="Channel", width=120, headerTooltip="Channel of the step (e.g. garden or grapher)."
-)
-gb.configure_column("namespace", headerName="Namespace", width=150, headerTooltip="Namespace of the step.")
-gb.configure_column("version", headerName="Version", width=120, headerTooltip="Version of the step.")
-gb.configure_column("name", headerName="Step name", width=140, headerTooltip="Short name of the step.")
-gb.configure_column("kind", headerName="Kind", width=100, headerTooltip="Kind of step (i.e. public or private).")
-gb.configure_column(
-    "n_charts", headerName="N. charts", width=120, headerTooltip="Number of charts that use data from the step."
-)
-# gb.configure_column(
-#     "n_chart_views_7d",
-#     headerName="7-day views",
-#     width=140,
-#     headerTooltip="Number of views of charts that use data from the step in the last 7 days.",
-# )
-gb.configure_column(
-    "n_chart_views_365d",
-    headerName="365-day views",
-    width=140,
-    headerTooltip="Number of views of charts that use data from the step in the last 365 days.",
-)
-gb.configure_column(
-    "date_of_next_update",
-    headerName="Next update",
-    width=140,
-    headerTooltip="Date of the next expected OWID update of the step.",
-)
-gb.configure_column(
-    "update_period_days",
-    headerName="Update period",
-    width=150,
-    headerTooltip="Number of days between consecutive OWID updates of the step.",
-)
-gb.configure_column(
-    "dag_file_name",
-    headerName="Name of DAG file",
-    width=160,
-    headerTooltip="Name of the DAG file that defines the step.",
-)
-gb.configure_column(
-    "full_path_to_script",
-    headerName="Path to script",
-    width=150,
-    headerTooltip="Path to the script that creates the ETL snapshot or dataset of this step.",
-)
-gb.configure_column(
-    "dag_file_path",
-    headerName="Path to DAG file",
-    width=160,
-    headerTooltip="Path to the DAG file that defines the step.",
-)
-gb.configure_column(
-    "n_versions",
-    headerName="N. versions",
-    width=140,
-    headerTooltip="Number of (active or archive) versions of the step.",
-)
-# Create a column with the number of days until the next expected update, colored according to its value.
-days_to_update_jscode = JsCode(
-    """
-    function(params){
-        if (params.value <= 0) {
-            return {
-                'color': 'black',
-                'backgroundColor': 'red'
-            }
-        } else if (params.value > 0 && params.value <= 31) {
-            return {
-                'color': 'black',
-                'backgroundColor': 'orange'
-            }
-        } else if (params.value > 31) {
-            return {
-                'color': 'black',
-                'backgroundColor': 'green'
-            }
-        } else {
-            return {
-                'color': 'black',
-                'backgroundColor': 'yellow'
-            }
-        }
-    }
-    """
-)
-gb.configure_columns(
-    "days_to_update",
-    cellStyle=days_to_update_jscode,
-    headerName="Days to update",
-    width=120,
-    headerTooltip="Number of days until the next expected OWID update of the step (if negative, an update is due).",
-)
-# Create a column colored depending on the update state.
-update_state_jscode = JsCode(
-    f"""
-function(params){{
-    if (params.value === "{UpdateState.UP_TO_DATE.value}") {{
-        return {{'color': 'black', 'backgroundColor': 'green'}}
-    }} else if (params.value === "{UpdateState.OUTDATED.value}") {{
-        return {{'color': 'black', 'backgroundColor': 'gray'}}
-    }} else if (params.value === "{UpdateState.MAJOR_UPDATE.value}") {{
-        return {{'color': 'black', 'backgroundColor': 'red'}}
-    }} else if (params.value === "{UpdateState.MINOR_UPDATE.value}") {{
-        return {{'color': 'black', 'backgroundColor': 'orange'}}
-    }} else if (params.value === "{UpdateState.ARCHIVABLE.value}") {{
-        return {{'color': 'black', 'backgroundColor': 'blue'}}
-    }} else if (params.value === "{UpdateState.UNUSED.value}") {{
-        return {{'color': 'black', 'backgroundColor': 'lightblue'}}
-    }} else {{
-        return {{'color': 'black', 'backgroundColor': 'yellow'}}
-    }}
-}}
-"""
-)
-gb.configure_columns(
-    "update_state",
-    headerName="Update state",
-    cellStyle=update_state_jscode,
-    width=150,
-    headerTooltip=f'Update state of the step: "{UpdateState.UP_TO_DATE.value}" (up to date), "{UpdateState.MINOR_UPDATE.value}" (a dependency is outdated, but all origins are up to date), "{UpdateState.MAJOR_UPDATE.value}" (an origin is outdated), "{UpdateState.OUTDATED.value}" (there is a newer version of the step), "{UpdateState.ARCHIVABLE.value}" (the step is outdated and not used in charts, therefore can safely be archived).',
-)
-# Create a column with grapher dataset names that are clickable and open in a new tab.
-grapher_dataset_jscode = JsCode(
-    r"""
-    class UrlCellRenderer {
-    init(params) {
-        this.eGui = document.createElement('a');
-        if(params.value) {
-            const match = params.value.match(/\[(.*?)\]\((.*?)\)/);
-            if(match && match.length >= 3) {
-                const datasetName = match[1];
-                const datasetUrl = match[2];
-                this.eGui.innerText = datasetName;
-                this.eGui.setAttribute('href', datasetUrl);
-            } else {
-                this.eGui.innerText = '';
-            }
-        } else {
-            this.eGui.innerText = '';
-        }
-        this.eGui.setAttribute('style', "text-decoration:none");
-        this.eGui.setAttribute('target', "_blank");
-    }
-    getGui() {
-        return this.eGui;
-    }
-    }
-    """
-)
-gb.configure_column(
-    "db_dataset_name_and_url",  # This will be the displayed column
-    headerName="Grapher dataset",
-    cellRenderer=grapher_dataset_jscode,
-    headerTooltip="Name of the grapher dataset (if any), linked to its corresponding dataset admin page.",
-)
-gb.configure_pagination(paginationAutoPageSize=False, paginationPageSize=20)
-grid_options = gb.build()
-
-# Display the grid table with pagination.
-grid_response = AgGrid(
-    data=steps_df_display,
-    gridOptions=grid_options,
-    height=1000,
-    width="100%",
-    update_mode=GridUpdateMode.MODEL_CHANGED,
-    fit_columns_on_grid_load=False,
-    allow_unsafe_jscode=True,
-    theme="streamlit",
-    # The following ensures that the pagination controls are not cropped.
-    custom_css={
-        "#gridToolBar": {
-            "padding-bottom": "0px !important",
-        }
-    },
-)
+# Build and display the grid table with pagination.
+grid_response = make_agrid(steps_df_display)
 
 ########################################
 # OPERATIONS LIST MANAGEMENT
